refactor: 입력된 로또 번호 검증로직 Lotto 클래스로 이동.
- 수동 로또를 추가하면서 CorrectNumbers 객체를 생성하지 않고, 로또 객체를 생성할 때 검증하도록 변경.
